Updates to try to get tensorflow classifier working

+++ PyTorch2Java/jdl4pam/src/main/java/org/jamdev/jdl4pam/genericmodel/GenericModel.java
@@ -37,7 +37,7 @@ public class GenericModel {
 	/**
 	 * The predictor for the model. 
 	 */
-	Predictor<double[][], float[]> predictor; 
+	Predictor<float[][], float[]> predictor; 
 
 
 	public GenericModel(String modelPath) throws MalformedModelException, IOException{
@@ -67,7 +67,7 @@ public class GenericModel {
 	 * @param specImage - the spectrogram image
 	 * @return the results 
 	 */
-	public float[] runModel(double[][] specImage) {
+	public float[] runModel(float[][] specImage) {
 		try {
 			float[] results  = predictor.predict(specImage);
 			//DLUtils.printArray(results);
